requirements.txt
opencv-python
scikit-learn
pandas
numpy 
matplotlib
pyyaml
